system - allow installing of retropie on rpi ubuntu (source only).

+++ scriptmodules/system.sh
@@ -79,11 +79,10 @@ function get_os_version() {
         else
             case "$ver" in
                 jessie/sid|stretch/sid)
-                    if ! isPlatform "rpi"; then
-                        __raspbian_ver=8
-                        __raspbian_name="ubuntu"
-                    else
-                        fatalError "Ubuntu is not supported on the Raspberry Pi - Please use Raspbian."
+                    __raspbian_ver=8
+                    __raspbian_name="ubuntu"
+                    if isPlatform "rpi"; then
+                        __has_binaries=0
                     fi
                     return
                     ;;
@@ -133,7 +132,7 @@ function set_default_gcc() {
 function get_retropie_depends() {
     # add rasberrypi repository if it's missing (needed for libraspberrypi-dev etc) - not used on osmc
     local config="/etc/apt/sources.list.d/raspi.list"
-    if [[ ! -f "$config" ]] && hasPackage raspberrypi-bootloader; then
+    if [[ ! -f "$config" && "$__raspbian_name" != "ubuntu" ]] && hasPackage raspberrypi-bootloader; then
         # add key
         wget -q http://archive.raspberrypi.org/debian/raspberrypi.gpg.key -O- | apt-key add - >/dev/null
         echo "deb http://archive.raspberrypi.org/debian/ $__raspbian_name main" >>$config
